src/requirements.txt
asgiref==3.2.10
django==3.1.1
django-k8s==0.2.9
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
mysqlclient==2.0.1
psycopg2-binary==2.8.6
pyjwt==1.7.1
pytz==2020.1
sqlparse==0.3.1
